Создал синглтон, немного оптимизации
