+++ vkApp/Controller/FriendList/FriendListCell.swift
