+++ vkApp/Controller/FriendList/FriendListTableController.swift
@@ -1,10 +1,3 @@
-//
-//  FriendListTableViewController.swift
-//  vkApp
-//
-//  Created by Никита Ананьев on 02.04.2020.
-//  Copyright © 2020 @NkitaAnanev. All rights reserved.
-//
 
 import UIKit
