+++ vkApp/Controller/FriendPhoto/FriendPhotoCell.swift
