+++ vkApp/Controller/FriendPhoto/FriendPhotoCollectionController.swift
