+++ vkApp/Controller/News/NewsCollectionViewController.swift
