+++ vkApp/Controller/News/NewsCollectionViewLayout.swift
