+++ vkApp/Controller/SearchCommunity/SearchComminityCell.swift
