+++ vkApp/Controller/SearchCommunity/shadowView.swift
