+++ vkApp/Delegate/AppDelegate.swift
