+++ vkApp/LaunchScreenViewController.swift
@@ -1,0 +1,17 @@
+//
+//  LaunchScreenViewController.swift
+//  vkApp
+//
+//  Created by Никита Ананьев on 21.04.2020.
+//  Copyright © 2020 @NkitaAnanev. All rights reserved.
+//
+
+import UIKit
+
+class LaunchScreenViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+}
