+++ vkApp/Model/Community.swift
