+++ vkApp/Model/ShadowView.swift
